Improve TagLayoutView for colorScheme dark and light
